Create bitwarden-test and refactor code
