+++ crates/bitwarden-test/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "bitwarden-test"
+description = """
+Internal crate for the bitwarden crate. Do not use.
+"""
+
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+readme.workspace = true
+homepage.workspace = true
+repository.workspace = true
+license-file.workspace = true
+keywords.workspace = true
+
+[dependencies]
+async-trait = { workspace = true }
+bitwarden-api-api.workspace = true
+bitwarden-state = { workspace = true }
+reqwest = { workspace = true }
+wiremock = { workspace = true }
+
+[lints]
+workspace = true
