+++ crates/bitwarden-test/README.md
@@ -1,0 +1,7 @@
+# Bitwarden Test
+
+<div class="warning">
+This crate should only be used in tests and should not be included in production builds.
+</div>
+
+Contains test utilities for Bitwarden.
