+++ crates/bitwarden-test/src/lib.rs
@@ -1,0 +1,7 @@
+#![doc = include_str!("../README.md")]
+
+mod api;
+pub use api::*;
+
+mod repository;
+pub use repository::*;
