+++ crates/bitwarden-vault/src/folder/folder_create.rs
@@ -1,5 +1,3 @@
-use std::sync::Arc;
-
 use bitwarden_api_api::{apis::folders_api, models::FolderRequestModel};
 use bitwarden_core::{
     key_management::{KeyIds, SymmetricKeyId},
@@ -66,7 +64,7 @@ pub enum CreateFolderError {
 pub(super) async fn create_folder<R: Repository<Folder> + ?Sized>(
     key_store: &KeyStore<KeyIds>,
     api_config: &bitwarden_api_api::apis::configuration::Configuration,
-    repository: &Arc<R>,
+    repository: &R,
     request: FolderAddEditRequest,
 ) -> Result<FolderView, CreateFolderError> {
     let folder_request = key_store.encrypt(request)?;
@@ -87,9 +85,9 @@ pub(super) async fn create_folder<R: Repository<Folder> + ?Sized>(
 mod tests {
     use bitwarden_api_api::models::FolderResponseModel;
     use bitwarden_crypto::SymmetricCryptoKey;
-    use bitwarden_state::repository::MemoryRepository;
+    use bitwarden_test::{start_api_mock, MemoryRepository};
     use uuid::uuid;
-    use wiremock::{matchers, Mock, MockServer, Request, ResponseTemplate};
+    use wiremock::{matchers, Mock, Request, ResponseTemplate};
 
     use super::*;
 
@@ -102,40 +100,31 @@ mod tests {
             SymmetricCryptoKey::make_aes256_cbc_hmac_key(),
         );
 
-        let server = MockServer::start().await;
-        server
-            .register(
-                Mock::given(matchers::path("/folders"))
-                    .respond_with(|req: &Request| {
-                        let body: FolderRequestModel = req.body_json().unwrap();
-                        ResponseTemplate::new(201).set_body_json(FolderResponseModel {
-                            id: Some(uuid!("25afb11c-9c95-4db5-8bac-c21cb204a3f1")),
-                            name: Some(body.name),
-                            revision_date: Some("2025-01-01T00:00:00Z".to_string()),
-                            object: Some("folder".to_string()),
-                        })
-                    })
-                    .expect(1),
-            )
-            .await;
+        let (_server, api_config) = start_api_mock(vec![Mock::given(matchers::path("/folders"))
+            .respond_with(|req: &Request| {
+                let body: FolderRequestModel = req.body_json().unwrap();
+                ResponseTemplate::new(201).set_body_json(FolderResponseModel {
+                    id: Some(uuid!("25afb11c-9c95-4db5-8bac-c21cb204a3f1")),
+                    name: Some(body.name),
+                    revision_date: Some("2025-01-01T00:00:00Z".to_string()),
+                    object: Some("folder".to_string()),
+                })
+            })
+            .expect(1)])
+        .await;
 
-        let request = FolderAddEditRequest {
-            name: "test".to_string(),
-        };
-        let api_config = &bitwarden_api_api::apis::configuration::Configuration {
-            base_path: server.uri(),
-            user_agent: Some("test-agent".to_string()),
-            client: reqwest::Client::new(),
-            basic_auth: None,
-            oauth_access_token: None,
-            bearer_access_token: None,
-            api_key: None,
-        };
-        let repository = Arc::new(MemoryRepository::<Folder>::new());
+        let repository = MemoryRepository::<Folder>::default();
 
-        let result = create_folder(&store, api_config, &repository, request)
-            .await
-            .unwrap();
+        let result = create_folder(
+            &store,
+            &api_config,
+            &repository,
+            FolderAddEditRequest {
+                name: "test".to_string(),
+            },
+        )
+        .await
+        .unwrap();
 
         assert_eq!(
             result,
