+++ crates/bitwarden-vault/src/folder/folder_edit.rs
@@ -1,5 +1,3 @@
-use std::sync::Arc;
-
 use bitwarden_api_api::apis::folders_api;
 use bitwarden_core::{key_management::KeyIds, require, ApiError, MissingFieldError};
 use bitwarden_crypto::{CryptoError, KeyStore};
@@ -32,11 +30,11 @@ pub enum EditFolderError {
 pub(super) async fn edit_folder<R: Repository<Folder> + ?Sized>(
     key_store: &KeyStore<KeyIds>,
     api_config: &bitwarden_api_api::apis::configuration::Configuration,
-    repository: &Arc<R>,
+    repository: &R,
     folder_id: &str,
     request: FolderAddEditRequest,
 ) -> Result<FolderView, EditFolderError> {
-    // Check if the folder exists
+    // Verify the folder we're updating exists
     repository
         .get(folder_id.to_owned())
         .await?
@@ -62,9 +60,9 @@ mod tests {
     use bitwarden_api_api::models::{FolderRequestModel, FolderResponseModel};
     use bitwarden_core::key_management::SymmetricKeyId;
     use bitwarden_crypto::{PrimitiveEncryptable, SymmetricCryptoKey};
-    use bitwarden_state::repository::MemoryRepository;
+    use bitwarden_test::{start_api_mock, MemoryRepository};
     use uuid::uuid;
-    use wiremock::{matchers, Mock, MockServer, Request, ResponseTemplate};
+    use wiremock::{matchers, Mock, Request, ResponseTemplate};
 
     use super::*;
 
@@ -79,36 +77,23 @@ mod tests {
 
         let folder_id = uuid!("25afb11c-9c95-4db5-8bac-c21cb204a3f1");
 
-        let server = MockServer::start().await;
-        server
-            .register(
-                Mock::given(matchers::path(format!("/folders/{}", folder_id)))
-                    .respond_with(move |req: &Request| {
-                        let body: FolderRequestModel = req.body_json().unwrap();
-                        ResponseTemplate::new(201).set_body_json(FolderResponseModel {
-                            id: Some(folder_id),
-                            name: Some(body.name),
-                            revision_date: Some("2025-01-01T00:00:00Z".to_string()),
-                            object: Some("folder".to_string()),
-                        })
-                    })
-                    .expect(1),
-            )
-            .await;
+        let (_server, api_config) = start_api_mock(vec![Mock::given(matchers::path(format!(
+            "/folders/{}",
+            folder_id
+        )))
+        .respond_with(move |req: &Request| {
+            let body: FolderRequestModel = req.body_json().unwrap();
+            ResponseTemplate::new(200).set_body_json(FolderResponseModel {
+                object: Some("folder".to_string()),
+                id: Some(folder_id),
+                name: Some(body.name),
+                revision_date: Some("2025-01-01T00:00:00Z".to_string()),
+            })
+        })
+        .expect(1)])
+        .await;
 
-        let request = FolderAddEditRequest {
-            name: "test".to_string(),
-        };
-        let api_config = &bitwarden_api_api::apis::configuration::Configuration {
-            base_path: server.uri(),
-            user_agent: Some("test-agent".to_string()),
-            client: reqwest::Client::new(),
-            basic_auth: None,
-            oauth_access_token: None,
-            bearer_access_token: None,
-            api_key: None,
-        };
-        let repository = Arc::new(MemoryRepository::<Folder>::new());
+        let repository = MemoryRepository::<Folder>::default();
 
         repository
             .set(
@@ -116,7 +101,7 @@ mod tests {
                 Folder {
                     id: Some(folder_id),
                     name: "old_name"
-                        .encrypt(&mut store.context_mut(), SymmetricKeyId::User)
+                        .encrypt(&mut store.context(), SymmetricKeyId::User)
                         .unwrap(),
                     revision_date: "2024-01-01T00:00:00Z".parse().unwrap(),
                 },
@@ -126,10 +111,12 @@ mod tests {
 
         let result = edit_folder(
             &store,
-            api_config,
+            &api_config,
             &repository,
             &folder_id.to_string(),
-            request,
+            FolderAddEditRequest {
+                name: "test".to_string(),
+            },
         )
         .await
         .unwrap();
@@ -137,7 +124,7 @@ mod tests {
         assert_eq!(
             result,
             FolderView {
-                id: Some(uuid!("25afb11c-9c95-4db5-8bac-c21cb204a3f1")),
+                id: Some(folder_id),
                 name: "test".to_string(),
                 revision_date: "2025-01-01T00:00:00Z".parse().unwrap(),
             }
